dist/datachannel_test_2peer/create/003.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Create Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
           var label = "test";
            var dataChannel = null;
            var test = async_test("Set up a DataChannel - close RTCPeerConnection (remote peer closes RTCPeerConnection) - readyState on both peers should be closed");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    assert_equals(dataChannel.readyState, "connecting", "Wrong readyState");
                    dataChannel.onopen = test.step_func(function() {
						setTimeout(function() {
                            test.step(function() {
                                assert_equals(dataChannel.readyState, "closed", "Wrong readyState");
                                test.done();
                            });
                        }, 3000);
                    });
                });
            } else {
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;

                    dataChannel.onopen = test.step_func(function() {
                        setTimeout(test.step_func(function() {
                             peerConnection.close();
                        }), 1000);
                        setTimeout(test.step_func(function() {
                            assert_equals(dataChannel.readyState, "closed", "Wrong readyState");
                            test.done();
                        }), 3000);
                    });
                });
            }
		</script>
	</body>
</html>

